+++ app/src/main/java/in/testpress/testpress/core/TestpressService.java
@@ -189,7 +189,7 @@ public class TestpressService {
         return getPostService().sendComments(postId, params);
     }
 
-    public RegistrationSuccessResponse register(String username,String email, String password, String phone, String countryCode){
+    public RegistrationSuccessResponse registerDevice(String username, String email, String password, String phone, String countryCode){
         HashMap<String, String> userDetails = new HashMap<String, String>();
         userDetails.put("username", username);
         userDetails.put("email", email);
